--- a/.travis.yml
+++ b/.travis.yml
@@ -4,7 +4,7 @@ dist: trusty
 before_install:
     - sudo add-apt-repository -y ppa:ubuntu-toolchain-r/test
     - sudo apt-get update -qq
-    - sudo apt-get install --yes -qq gcc-6 g++-6
+    - sudo apt-get install --yes -qq gcc-6 g++-6 gdb
     - sudo apt-get install libssl-dev open-iscsi docker.io libjson-c-dev
     # use gcc-6 by default
     - sudo unlink /usr/bin/gcc && sudo ln -s /usr/bin/gcc-6 /usr/bin/gcc