--- a/src/istgt_integration.h
+++ b/src/istgt_integration.h
@@ -62,11 +62,12 @@ typedef struct replica_s {
 	int dont_free;
 
 	struct timespec create_time;
+
 	/*
-	 * this variable will be updated by only replica_thread,
-	 * so no lock required while updating this
+	 * Following variables should be updated with atomic operation only
 	 */
 	uint64_t replica_inflight_write_io_cnt;
+	uint64_t replica_inflight_read_io_cnt;
 
 	zvol_io_hdr_t *io_resp_hdr;	/* header recieved on data connection */
 	int io_state;			/* state of command on data connection */