… test_istgt.sh