Changes : 

- selecting healthy replica with minimal in-flight read IOs to serve read IOs. With this approach, Workload for the read IOs will be shared across all healthy replicas, which will reduce latency for the read IOs.

FIO latency-throughput results for the read IOS with/without these changes.
**Without these changes:** 
Latency test (Latency: 39.34 ms, Iops: 813)
Throughput test (Latency: 41.37 ms, Iops: 3092)

**With these changes:** 
Latency test (Latency: 31.46 ms, Iops: 1016)
Throughput test (Latency: 33.38 ms, Iops: 3831)

###### _"These results are taken on a virtual machine with a dummy replica(`replication_test`) which was modified to deliver latency ~3000 u-sec). The configuration used for the above test:_
> _ReplicationFactor : 3_
_ConsistencyFactor : 2_
_Test run time: 30 second_
_FIO Block size: 4k_
_IO-depth: 32_
_Thread: 1 (for latency test) 4(for throughput test)_

Signed-off-by: mayank <mayank.patel@cloudbyte.com>